schedules working to a reasonalbe degree
